updated zeolite example with restart abilities

+++ examples/run_zeolite.py
@@ -8,6 +8,7 @@ import numpy as np
 from ase.io import read
 
 import parsl
+from parsl.utils import get_all_checkpoints
 
 from psiflow.manager import Manager
 from psiflow.learning import RandomLearning, OnlineLearning
@@ -23,23 +24,29 @@ from psiflow.utils import get_parsl_config_from_file
 
 def get_context_and_manager(args):
     path_run = Path.cwd() / args.name
-    if path_run.is_dir():
-        shutil.rmtree(path_run)
-    path_run.mkdir()
+    if not args.restart:
+        if path_run.is_dir():
+            shutil.rmtree(path_run)
+        path_run.mkdir()
+    else:
+        assert path_run.is_dir()
     path_internal = path_run / 'parsl_internal'
     path_context  = path_run / 'context_dir'
     config = get_parsl_config_from_file(
             args.parsl_config,
             path_internal,
             )
+    config.initialize_logging = False
+    config.checkpoint_mode = 'task_exit'
+    if args.restart:
+        config.checkpoint_files = get_all_checkpoints(str(path_internal))
+        print('found {} checkpoint files'.format(len(config.checkpoint_files)))
     parsl.load(config)
     config.retries = args.retries
-    config.cache=True
-    config.initialize_logging = False
     context = ExecutionContext(config, path=path_context)
     context.register(ModelExecutionDefinition())
-    context.register(ReferenceExecutionDefinition(time_per_singlepoint=500))
-    context.register(TrainingExecutionDefinition(walltime=3600))
+    context.register(ReferenceExecutionDefinition(time_per_singlepoint=30))
+    context.register(TrainingExecutionDefinition())
 
     # setup manager for IO, wandb logging
     path_output  = path_run / 'output'
@@ -47,9 +54,11 @@ def get_context_and_manager(args):
             path_output,
             wandb_project='zeolite',
             wandb_group=args.name,
+            restart=args.restart,
             error_x_axis='CV', # plot errors w.r.t CV value
             )
     return context, manager
+
 
 
 def get_bias(context):
@@ -144,6 +153,7 @@ if __name__ == '__main__':
     parser.add_argument('--parsl-config', action='store')
     parser.add_argument('--name', action='store')
     parser.add_argument('--retries', action='store', default=1)
+    parser.add_argument('--restart', action='store_true', default=False)
     args = parser.parse_args()
 
     context, manager = get_context_and_manager(args)
